Add end to end training deprecation test, fix validation bug

+++ src/metatrain/share/base_hypers.py
@@ -117,7 +117,7 @@ class SphericalTargetIrrepsConfig(TypedDict):
     specified in ``TargetHypers``.
     """
     o3_lambda: int
-    o3_sigma: float
+    o3_sigma: int
 
 
 @with_config(ConfigDict(extra="forbid", strict=True))
